Code/Python/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
altair=3.2.0=py37_0
appnope=0.1.0=py37_1000
attrs=19.3.0=py_0
backcall=0.1.0=py_0
bleach=3.1.0=py_0
boost-cpp=1.68.0=h6f8c590_1000
branca=0.3.1=py_0
bson=0.5.8=py_0
bzip2=1.0.8=h01d97ff_1
ca-certificates=2019.9.11=hecc5488_0
cairo=1.16.0=h0ab9d94_1001
certifi=2019.9.11=py37_0
cffi=1.13.2=py37h33e799b_0
cfitsio=3.470=h84d2f63_2
chardet=3.0.4=py37_1003
click=7.0=py_0
click-plugins=1.1.1=py_0
cligj=0.5.0=py_0
cloudpickle=1.2.2=py_0
cryptography=2.8=py37hafa8578_0
curl=7.65.3=h22ea746_0
cycler=0.10.0=py_2
cytoolz=0.10.1=py37h0b31af3_0
dask-core=2.6.0=py_0
decorator=4.4.1=py_0
defusedxml=0.6.0=py_0
descartes=1.1.0=py_4
entrypoints=0.3=py37_1000
expat=2.2.5=h4a8c4bd_1004
fiona=1.8.9.post2=py37h238e834_0
folium=0.10.0=py_0
fontconfig=2.13.1=h1027ab8_1000
freetype=2.10.0=h24853df_1
freexl=1.0.5=h1de35cc_1002
gdal=2.4.2=py37h39889d8_5
geographiclib=1.50=py_0
geopandas=0.6.1=py_1
geopy=1.20.0=py_0
geos=3.7.2=h6de7cb9_2
geotiff=1.5.1=h83de174_2
gettext=0.19.8.1=h46ab8bc_1002
giflib=5.1.7=h01d97ff_1
glib=2.58.3=h9d45998_1002
hdf4=4.2.13=hf3c6af0_1002
hdf5=1.10.5=nompi_h3e39495_1104
icu=58.2=h0a44026_1000
idna=2.8=py37_1000
imageio=2.6.1=py37_0
importlib_metadata=0.23=py37_0
intel-openmp=2019.4=233
ipykernel=5.1.3=py37h5ca1d4c_0
ipython=7.9.0=py37h5ca1d4c_0
ipython_genutils=0.2.0=py_1
jedi=0.15.1=py37_0
jinja2=2.10.3=py_0
jpeg=9c=h1de35cc_1001
json-c=0.13.1=h1de35cc_1001
json5=0.8.5=py_0
jsonschema=3.1.1=py37_0
jupyter_client=5.3.3=py37_1
jupyter_core=4.5.0=py_0
jupyterlab=1.2.1=py_0
jupyterlab_server=1.0.6=py_0
kealib=1.4.10=h6659575_1005
kiwisolver=1.1.0=py37ha1b3eb9_0
krb5=1.16.3=hcfa6398_1001
libblas=3.8.0=14_openblas
libcblas=3.8.0=14_openblas
libcurl=7.65.3=h16faf7d_0
libcxx=9.0.0=h89e68fa_1
libdap4=3.20.4=habf5908_0
libedit=3.1.20170329=hcfe32e1_1001
libffi=3.2.1=h6de7cb9_1006
libgdal=2.4.2=he37cffe_5
libgfortran=4.0.0=2
libiconv=1.15=h01d97ff_1005
libkml=1.3.0=h71ee1b2_1009
liblapack=3.8.0=14_openblas
libnetcdf=4.6.2=h1a02027_1003
libopenblas=0.3.7=h4bb4525_3
libpng=1.6.37=h2573ce8_0
libpq=11.5=h756f0eb_1
libsodium=1.0.17=h01d97ff_0
libspatialindex=1.9.3=h4a8c4bd_1
libspatialite=4.3.0a=he369b6e_1029
libssh2=1.8.2=hcdc9a53_2
libtiff=4.1.0=h3527a1b_0
libxml2=2.9.9=hd80cff7_2
llvm-openmp=8.0.1=h770b8ee_0
lz4-c=1.8.3=h6de7cb9_1001
markupsafe=1.1.1=py37h0b31af3_0
matplotlib=3.1.2=py37_0
matplotlib-base=3.1.2=py37h11da6c2_0
mistune=0.8.4=py37h0b31af3_1000
mkl=2019.4=233
more-itertools=7.2.0=py_0
munch=2.5.0=py_0
nbconvert=5.6.1=py37_0
nbformat=4.4.0=py_1
ncurses=6.1=h0a44026_1002
networkx=2.4=py_0
ninja=1.9.0=py37h04f5b5a_0
notebook=6.0.1=py37_0
numpy=1.17.3=py37hde6bac1_0
olefile=0.46=py_0
openjpeg=2.3.1=hdc36067_1
openssl=1.1.1d=h0b31af3_0
osmnx=0.10=py_1
pandas=0.25.3=py37h4f17bb1_0
pandoc=2.7.3=0
pandocfilters=1.4.2=py_1
parso=0.5.1=py_0
patsy=0.5.1=py37_0
pcre=8.43=h4a8c4bd_0
pexpect=4.7.0=py37_0
pickleshare=0.7.5=py37_1000
pillow=6.2.1=py37hb6f49c9_0
pip=19.3.1=py37_0
pixman=0.38.0=h01d97ff_1003
poppler=0.67.0=hd5eb092_7
poppler-data=0.4.9=1
postgresql=11.5=h25afefd_1
proj4=6.1.0=h2cc77ee_2
prometheus_client=0.7.1=py_0
prompt_toolkit=2.0.10=py_0
ptyprocess=0.6.0=py_1001
pycparser=2.19=py37_1
pygments=2.4.2=py_0
pygsp=0.5.1=py_0
pyopenssl=19.0.0=py37_0
pyparsing=2.4.4=py_0
pyproj=2.2.1=py37h804dea5_0
pyrsistent=0.15.5=py37h0b31af3_0
pysocks=1.7.1=py37_0
python=3.7.3=h93065d6_1
python-dateutil=2.8.1=py_0
pytorch=1.3.1=py3.7_0
pytz=2019.3=py_0
pywavelets=1.1.1=py37h3b54f70_0
pyzmq=18.1.0=py37h4bf09a9_0
readline=8.0=hcfe32e1_0
requests=2.22.0=py37_1
rtree=0.8.3=py37h7b0cdae_1003
scikit-image=0.16.2=py37h4f17bb1_0
scipy=1.3.1=py37h7e0e109_2
seaborn=0.9.0=pyh91ea838_1
send2trash=1.5.0=py_0
setuptools=41.6.0=py37_1
shapely=1.6.4=py37h5c88e11_1006
six=1.13.0=py37_0
sqlite=3.30.1=h93121df_0
statsmodels=0.10.1=py37h1d22016_0
terminado=0.8.2=py37_0
testpath=0.4.4=py_0
tk=8.6.9=h2573ce8_1003
toolz=0.10.0=py_0
tornado=6.0.3=py37h0b31af3_0
tqdm=4.36.1=py_0
traitlets=4.3.3=py37_0
tzcode=2019a=h01d97ff_1002
urllib3=1.25.6=py37_0
vincent=0.4.4=py_1
wcwidth=0.1.7=py_1
webencodings=0.5.1=py_1
wheel=0.33.6=py37_0
xerces-c=3.2.2=h4db8090_1003
xz=5.2.4=h1de35cc_1001
zeromq=4.3.2=h6de7cb9_2
zipp=0.6.0=py_0
zlib=1.2.11=h0b31af3_1006
zstd=1.4.3=he7fca8b_0
